internal/collector: truncate timestamps in filter
Most of the Event's timestamps are in metav1.Time and are truncated to
the second. As we are comparing the Events latest timestamps to the
timestamp at which the exporter started running, we need to truncate
this timestamp in order to record the Events created during the same
second as the exporter.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>