test/framework: slightly increase default timeout
The TestEventUpdate test was failing from time to time. However, I could
not reproduce this failure and thus not assert if the issue was only
affecting this test or not. Looking into the code, this seems to be a
timeout issue that occurs very rarely so I slightly increase the
framework default timeout from 5 to 7 seconds.

Signed-off-by: Damien Grisonnet <dgrisonn@redhat.com>